+++ package.json
@@ -22,6 +22,8 @@
     "concurrently": "^5.3.0"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^1.2.32",
+    "@fortawesome/react-fontawesome": "^0.1.11",
     "dotenv": "^8.2.0"
   }
 }
