--- public/src/Footer/_footerVars.js
@@ -1,40 +1,0 @@
-let footerVars = {
-    navEnglish: [
-        {
-            name: 'Home',
-            link: '#',
-        },
-        {
-            name: 'Store',
-            link: '#',
-        },
-        {
-            name: 'F.A.Q.',
-            link: '#',
-        },
-        {
-            name: 'Our Facilities',
-            link: '#'
-        },
-        {
-            name: 'Contact Us',
-            link: '#',
-        },
-        {
-            name: 'Create Account',
-            link: '#',
-        },
-        {
-            name: 'Login',
-            link: '#',
-        },
-    ],
-    socialMedia: [
-        '#',
-        '#',
-        '#',
-        '#',
-    ],
-};
-
-export default footerVars;
